requirements.txt
polygon-api-client
mplfinance
pandas
matplotlib
requests
python-dotenv
Pillow
fastapi
uvicorn[standard]
pywebpush
